(maint) Organize mock files
